cast every thing
Former-commit-id: 4f3810dd92ae0acacb01af5cd9b23d7b2b9d3f57
Former-commit-id: 17de296b984c8a02c3a383b35e68c5927adeafc9

+++ JAxtar/annotate.py
@@ -3,5 +3,5 @@ import jax.numpy as jnp
 KEY_DTYPE = jnp.float16
 ACTION_DTYPE = jnp.uint8
 HASH_POINT_DTYPE = jnp.uint32
-HASH_TABLE_IDX_DTYPE = jnp.uint32
+HASH_TABLE_IDX_DTYPE = jnp.uint8
 SIZE_DTYPE = jnp.uint32
